ANTLR util refactor
